No windows.h in header file
